+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * Readme instructions to install from PyPI.
+* Enabled directory listing.
 
 ## v0.1.0 (2019-06-13)